requirements.txt
# Database and ORM
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Optional (if you want async support later)
asyncpg==0.29.0
pandas==2.2.3
numpy==1.26.4

fastapi[standard]
alembic==1.16.5
geoalchemy==2-0.18.0 